Refactor admin module
