+++ lib/auth.js
@@ -59,7 +59,6 @@ export async function isAuthenticated(ctx) {
     // Next check the cookie
     const cookie = nextCookies(ctx)
     user = cookie.user
-    user.adminTempAccessGroup = cookie.adminTempAccessGroup
 
     // Ensure user exists and is valid
     if (!user || typeof user.expiresAt !== 'number')
@@ -71,6 +70,10 @@ export async function isAuthenticated(ctx) {
       user = await renewSession()
     }
 
+    // Set the adminTempAccessGroup from the cookie
+    user.adminTempAccessGroup = cookie.adminTempAccessGroup
+
+    // Update the Redux store with the user info
     ctx.store.dispatch(setUser(user))
   }
