--- lib/modules/admin/components/index.js
@@ -1,3 +1,0 @@
-export {default as JobDashboard} from './job-dashboard'
-export {default as MainDashboard} from './main-dashboard'
-export {default as WorkerDashboard} from './worker-dashboard'
