--- lib/modules/admin/constants.js
@@ -1,3 +1,0 @@
-//
-export const REFRESH_INTERVAL_MS = 5000
-export const START_TIME_FORMAT = 'MM-dd HH:mm:ss'
