--- lib/modules/admin/jobs.json
@@ -1,113 +1,0 @@
-[
-  {
-    "jobId": "5dee381f6844a22caafc8551",
-    "graphId": "5d280afc32b98e765c94557a",
-    "workerCommit": "v4.7.0",
-    "total": 7392,
-    "complete": 1064,
-    "incomplete": 6328,
-    "deliveries": 1104,
-    "deliveryPass": 0,
-    "regionalAnalysis": {
-      "_id": "5dee381f6844a22caafc8551",
-      "name": "Analysis 5: 2019-07-12 All Modifications ",
-      "nonce": "5dee381f6844a22caafc8552",
-      "createdAt": 1575893023260,
-      "updatedAt": 1575893023260,
-      "accessGroup": "conveyal",
-      "createdBy": "trevorgerhardt@gmail.com",
-      "updatedBy": "trevorgerhardt@gmail.com",
-      "regionId": "5d280ab432b98e765c945576",
-      "bundleId": "5d280afc32b98e765c94557a",
-      "projectId": "5d280bd432b98e765c94557d",
-      "variant": -1,
-      "workerVersion": "v4.7.0",
-      "zoom": 9,
-      "width": 77,
-      "height": 96,
-      "north": 50095,
-      "west": 37458,
-      "request": {
-        "type": "REGIONAL_ANALYSIS",
-        "fromLat": 38.8823127746582,
-        "fromLon": -77.01683044433594,
-        "toLat": 0,
-        "toLon": 0,
-        "fromTime": 28800,
-        "toTime": 32400,
-        "walkSpeed": 1.3888888,
-        "bikeSpeed": 4.1666665,
-        "bikeTrafficStress": 4,
-        "carSpeed": 20,
-        "streetTime": 90,
-        "maxWalkTime": 20,
-        "maxBikeTime": 20,
-        "maxCarTime": 45,
-        "minBikeTime": 10,
-        "minCarTime": 10,
-        "date": "2019-08-05",
-        "limit": 0,
-        "accessModes": "WALK",
-        "egressModes": "WALK",
-        "directModes": "WALK",
-        "transitModes": "TRAM,SUBWAY,RAIL,BUS,FERRY,CABLE_CAR,GONDOLA,FUNICULAR",
-        "suboptimalMinutes": 5,
-        "maxTripDurationMinutes": 30,
-        "maxRides": 4,
-        "scenario": {
-          "id": "5d280bd432b98e765c94557d--1-223132457",
-          "description": "no description provided",
-          "modifications": [],
-          "feedChecksums": null
-        },
-        "scenarioId": "5d280bd432b98e765c94557d--1-223132457",
-        "zoneId": "Z",
-        "wheelchair": false,
-        "maxFare": 0,
-        "monteCarloDraws": 1,
-        "zoom": 9,
-        "west": 37458,
-        "north": 50095,
-        "width": 77,
-        "height": 96,
-        "graphId": "5d280afc32b98e765c94557a",
-        "workerVersion": "v4.7.0",
-        "jobId": "5d280bd432b98e765c94557d--1-223132457",
-        "originId": null,
-        "taskId": 0,
-        "makeTauiSite": false,
-        "computeTravelTimeBreakdown": false,
-        "computePaths": false,
-        "percentiles": [50],
-        "nPathsPerTarget": 3,
-        "logRequest": false,
-        "grid": "5d280ab432b98e765c945576/5d28606532b98e765c94559a.grid",
-        "originPointSetKey": null,
-        "oneToOne": false,
-        "recordTimes": false,
-        "recordAccessibility": true
-      },
-      "travelTimePercentile": 50,
-      "grid": "5d28606532b98e765c94559a",
-      "cutoffMinutes": 30,
-      "bounds": null,
-      "complete": false,
-      "deleted": false,
-      "status": {
-        "total": 7392,
-        "complete": 1064
-      }
-    }
-  },
-  {
-    "jobId": "SUM",
-    "graphId": "SUM",
-    "workerCommit": null,
-    "total": 7392,
-    "complete": 1064,
-    "incomplete": 6328,
-    "deliveries": 0,
-    "deliveryPass": 0,
-    "regionalAnalysis": null
-  }
-]
