--- lib/modules/admin/types.js
@@ -1,1 +1,0 @@
-//
