--- lib/modules/admin/workers.json
@@ -1,200 +1,0 @@
-[
-  {
-    "workerId": "eb238715bef04d049534ddd507a85c11",
-    "architecture": "amd64",
-    "processors": 4,
-    "loadAverage": 0,
-    "osName": "Linux",
-    "osVersion": "4.9.17-8.31.amzn1.x86_64",
-    "memoryMax": 27653046272,
-    "memoryTotal": 1449132032,
-    "memoryFree": 550524224,
-    "workerName": "R5",
-    "workerVersion": "v4.7.0",
-    "networks": ["5d280afc32b98e765c94557a"],
-    "scenarios": ["5d280bd432b98e765c94557d--1-223132457"],
-    "secondsSinceLastPoll": 16.004,
-    "tasksPerMinuteByJobId": {},
-    "ec2region": "eu-west-1",
-    "ec2instanceId": "i-0d9f61e0d29e014a9",
-    "ec2instanceType": "r5.xlarge",
-    "ec2machineImage": "ami-e5083683",
-    "ec2privateIp": "10.0.0.43",
-    "jvmStartTime": 1575892569,
-    "jvmUptime": 321,
-    "jvmName": "OpenJDK 64-Bit Server VM",
-    "jvmVendor": "Oracle Corporation",
-    "jvmVersion": "25.222-b10",
-    "ipAddress": "10.0.0.43",
-    "results": null,
-    "bundles": [
-      {
-        "_id": "5d280afc32b98e765c94557a",
-        "name": "WMATA",
-        "nonce": "5d280b6032b98e765c94557c",
-        "createdAt": 1562905340435,
-        "updatedAt": 1562905440022,
-        "accessGroup": "conveyal",
-        "createdBy": "trevorgerhardt@gmail.com",
-        "updatedBy": "trevorgerhardt@gmail.com",
-        "regionId": "5d280ab432b98e765c945576",
-        "north": 39.191489,
-        "south": 38.694261,
-        "east": -76.668939,
-        "west": -77.449457,
-        "serviceStart": "2019-04-10",
-        "serviceEnd": "2019-10-06",
-        "feeds": [
-          {
-            "feedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212",
-            "name": "WMATA",
-            "originalFileName": null,
-            "fileName": null,
-            "bundleScopedFeedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212_5d280afc32b98e765c94557a",
-            "serviceStart": "2019-04-10",
-            "serviceEnd": "2019-10-06",
-            "checksum": 4290348123
-          }
-        ],
-        "status": "DONE",
-        "feedsComplete": 1,
-        "totalFeeds": 1,
-        "errorCode": null
-      }
-    ],
-    "seenSecondsAgo": 1,
-    "lastSeenAt": 1575892890259
-  },
-  {
-    "workerId": "b385fda6afaa4a0cbbc3d388485e9d99",
-    "architecture": "amd64",
-    "processors": 4,
-    "loadAverage": 0,
-    "osName": "Linux",
-    "osVersion": "4.9.17-8.31.amzn1.x86_64",
-    "memoryMax": 27653046272,
-    "memoryTotal": 1439694848,
-    "memoryFree": 856287120,
-    "workerName": "R5",
-    "workerVersion": "v4.6.0-21-gd6b8630",
-    "networks": ["5d280afc32b98e765c94557a"],
-    "scenarios": ["5d280bd432b98e765c94557d--1-223132457"],
-    "secondsSinceLastPoll": 17.003,
-    "tasksPerMinuteByJobId": {},
-    "ec2region": "eu-west-1",
-    "ec2instanceId": "i-00bef2499bc87d97b",
-    "ec2instanceType": "r5.xlarge",
-    "ec2machineImage": "ami-e5083683",
-    "ec2privateIp": "10.0.0.229",
-    "jvmStartTime": 1575891783,
-    "jvmUptime": 1094,
-    "jvmName": "OpenJDK 64-Bit Server VM",
-    "jvmVendor": "Oracle Corporation",
-    "jvmVersion": "25.222-b10",
-    "ipAddress": "10.0.0.229",
-    "results": null,
-    "bundles": [
-      {
-        "_id": "5d280afc32b98e765c94557a",
-        "name": "WMATA",
-        "nonce": "5d280b6032b98e765c94557c",
-        "createdAt": 1562905340435,
-        "updatedAt": 1562905440022,
-        "accessGroup": "conveyal",
-        "createdBy": "trevorgerhardt@gmail.com",
-        "updatedBy": "trevorgerhardt@gmail.com",
-        "regionId": "5d280ab432b98e765c945576",
-        "north": 39.191489,
-        "south": 38.694261,
-        "east": -76.668939,
-        "west": -77.449457,
-        "serviceStart": "2019-04-10",
-        "serviceEnd": "2019-10-06",
-        "feeds": [
-          {
-            "feedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212",
-            "name": "WMATA",
-            "originalFileName": null,
-            "fileName": null,
-            "bundleScopedFeedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212_5d280afc32b98e765c94557a",
-            "serviceStart": "2019-04-10",
-            "serviceEnd": "2019-10-06",
-            "checksum": 4290348123
-          }
-        ],
-        "status": "DONE",
-        "feedsComplete": 1,
-        "totalFeeds": 1,
-        "errorCode": null
-      }
-    ],
-    "seenSecondsAgo": 13,
-    "lastSeenAt": 1575892878259
-  },
-  {
-    "workerId": "530f3c85e6374bf79e3dec1fd8b82a9b",
-    "architecture": "amd64",
-    "processors": 4,
-    "loadAverage": 0,
-    "osName": "Linux",
-    "osVersion": "4.9.17-8.31.amzn1.x86_64",
-    "memoryMax": 27653046272,
-    "memoryTotal": 8408006656,
-    "memoryFree": 3386243896,
-    "workerName": "R5",
-    "workerVersion": "v4.7.1-121-g7bc873a",
-    "networks": ["5d280afc32b98e765c94557a"],
-    "scenarios": [],
-    "secondsSinceLastPoll": 18.002,
-    "tasksPerMinuteByJobId": {},
-    "ec2region": "eu-west-1",
-    "ec2instanceId": "i-0dfb649e46b2ad50e",
-    "ec2instanceType": "r5.xlarge",
-    "ec2machineImage": "ami-e5083683",
-    "ec2privateIp": "10.0.0.136",
-    "jvmStartTime": 1575891948,
-    "jvmUptime": 936,
-    "jvmName": "OpenJDK 64-Bit Server VM",
-    "jvmVendor": "Oracle Corporation",
-    "jvmVersion": "25.222-b10",
-    "ipAddress": "10.0.0.136",
-    "results": null,
-    "bundles": [
-      {
-        "_id": "5d280afc32b98e765c94557a",
-        "name": "WMATA",
-        "nonce": "5d280b6032b98e765c94557c",
-        "createdAt": 1562905340435,
-        "updatedAt": 1562905440022,
-        "accessGroup": "conveyal",
-        "createdBy": "trevorgerhardt@gmail.com",
-        "updatedBy": "trevorgerhardt@gmail.com",
-        "regionId": "5d280ab432b98e765c945576",
-        "north": 39.191489,
-        "south": 38.694261,
-        "east": -76.668939,
-        "west": -77.449457,
-        "serviceStart": "2019-04-10",
-        "serviceEnd": "2019-10-06",
-        "feeds": [
-          {
-            "feedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212",
-            "name": "WMATA",
-            "originalFileName": null,
-            "fileName": null,
-            "bundleScopedFeedId": "0027a76b-cbe6-4e1d-9b14-453b259d1212_5d280afc32b98e765c94557a",
-            "serviceStart": "2019-04-10",
-            "serviceEnd": "2019-10-06",
-            "checksum": 4290348123
-          }
-        ],
-        "status": "DONE",
-        "feedsComplete": 1,
-        "totalFeeds": 1,
-        "errorCode": null
-      }
-    ],
-    "seenSecondsAgo": 6,
-    "lastSeenAt": 1575892885259
-  }
-]
