+++ lib/reducers.js
@@ -2,7 +2,6 @@ import unionBy from 'lodash/unionBy'
 import sortBy from 'lodash/sortBy'
 import {handleActions} from 'redux-actions'
 
-import admin from './modules/admin/reducer'
 import opportunityDatasets from './modules/opportunity-datasets/reducer'
 import r5Version from './modules/r5-version/reducer'
 import * as analysis from './reducers/analysis'
@@ -51,7 +50,6 @@ function resources(state = [], action) {
 }
 
 export default {
-  admin,
   analysis: handleActions(analysis.reducers, analysis.initialState),
   network: handleActions(network.reducers, network.initialState),
   opportunityDatasets,
