+++ package.json
@@ -77,6 +77,7 @@
     "date-fns": "^2.8.1",
     "debug": "^4.1.1",
     "emotion-theming": "^10.0.19",
+    "http-proxy": "^1.18.0",
     "js-cookie": "^2.2.1",
     "jsolines": "^2.1.2",
     "leaflet": "^1.6.0",
